requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
chromadb
anthropic
python-dotenv
pdfplumber
python-docx
openpyxl
pytesseract
pdf2image
Pillow
python-jose[cryptography]
